+++ wasatch/ROI.py
@@ -1,3 +1,7 @@
+import logging
+
+log = logging.getLogger(__name__)
+
 class ROI:
     """
     This class encapsulates a Region Of Interest, which may be either horizontal 
@@ -8,18 +12,22 @@ class ROI:
     """
 
     def __init__(self, start, end):
-        self.start = start
-        self.end = end
+        self.start = int(start)
+        self.end = int(end)
         self.len = end - start + 1
 
     def valid(self):
         return self.start >= 0 and self.start < self.end
 
     def crop(self, spectrum):
-        return spectrum[self.start:self.end+1]
+        try:
+            return spectrum[self.start:self.end+1]
+        except:
+            log.error(f"unable to crop spectrum of len {len(spectrum)} to roi {self}", exc_info=1)
+            return 
 
     def contains(self, value):
         return self.start <= value <= self.end
 
-    def __str__(self):
+    def __repr__(self):
         return f"({self.start}, {self.end}) inclusive"