+++ wasatch/WasatchDeviceWrapper.py
@@ -130,11 +130,11 @@ class WasatchDeviceWrapper:
 
     ##
     # Instantiated by Controller.connect_new(), if-and-only-if a WasatchBus
-    # reports a DeviceID which has not already connected to the GUI.  The DeviceID
+    # reports a DeviceID which has not already connected to the GUI. The DeviceID
     # is "unique and relevant" to the bus which reported it, but neither the
-    # bus class nor instance is passed in to this object.  If the DeviceID looks like
-    # "USB:VID:PID:bus:addr", then it is presumably USB.  Future DeviceID formats 
-    # could include "FILE:/path/to/dir", etc.  However, device_id is just
+    # bus class nor instance is passed in to this object. If the DeviceID looks like
+    # "USB:VID:PID:bus:addr", then it is presumably USB. Future DeviceID formats 
+    # could include "FILE:/path/to/dir", etc. However, device_id is just
     # a string scalar to this class, and actually parsing / using it should be
     # entirely encapsulated within WasatchDevice and lower using DeviceID.
     def __init__(self, device_id, log_level, callback=None):
@@ -156,6 +156,7 @@ class WasatchDeviceWrapper:
         self.is_spi       = '0x0403' in str(device_id)
         self.mock         = 'MOCK' in str(device_id).upper()
         self.is_ble       = 'BLE' in str(device_id)
+        self.is_tcp       = 'TCP' in str(device_id)
         self.wrapper_worker = None
         self.connect_start_time = datetime.datetime(year=datetime.MAXYEAR, month=1, day=1)
 
@@ -225,13 +226,14 @@ class WasatchDeviceWrapper:
             is_andor       = self.is_andor,
             is_spi         = self.is_spi,
             is_ble         = self.is_ble,
+            is_tcp         = self.is_tcp,
             log_level      = self.log_level,
             callback       = self.callback)
         log.debug("device wrapper: Instance created for worker")
 
         self.wrapper_worker.daemon = True
-        log.debug("deivce wrapper: Initiating wrapper thread")
 
+        log.debug("deivce wrapper: starting WrapperWorker thread")
         self.wrapper_worker.start()
 
         # expect to read a single post-initialization SpectrometerSettings object off the queue
@@ -253,7 +255,10 @@ class WasatchDeviceWrapper:
             time.sleep(0.1)
 
     def poll_settings(self): 
-        """ @returns SpectrometerResponse(True) on success, (False) otherwise """
+        """ 
+        @returns SpectrometerResponse(True) on success, (False) otherwise 
+        @note this doesn't have a timeout -- that's in enlighten.Controller.check_ready_initialize
+        """
         log.debug("polling device settings")
         if not self.settings_queue.empty():
             result = self.settings_queue.get_nowait()